fold SolidDefinition into OpDefinition
