+++ python_modules/dagster/dagster/_core/definitions/op_definition.py
@@ -5,9 +5,12 @@ from typing import (
     Any,
     Callable,
     List,
+    Iterator,
     Mapping,
     Optional,
     Sequence,
+    Tuple,
+    TypeVar,
     Union,
     cast,
 )
@@ -17,25 +20,42 @@ from typing_extensions import TypeAlias, get_origin
 import dagster._check as check
 from dagster._annotations import public
 from dagster._config.config_schema import UserConfigSchema
+from dagster._core.decorator_utils import get_function_params
+from dagster._core.definitions.dependency import NodeHandle
+from dagster._core.definitions.node_definition import NodeDefinition
 from dagster._core.definitions.policy import RetryPolicy
 from dagster._core.errors import DagsterInvariantViolationError
 from dagster._utils.backcompat import canonicalize_backcompat_args, deprecation_warning
+from dagster._core.definitions.resource_requirement import (
+    InputManagerRequirement,
+    OpDefinitionResourceRequirement,
+    OutputManagerRequirement,
+    ResourceRequirement,
+)
+from dagster._core.definitions.solid_invocation import solid_invocation_result
+from dagster._core.errors import DagsterInvalidInvocationError, DagsterInvariantViolationError
+from dagster._core.types.dagster_type import DagsterType, DagsterTypeKind
+from dagster._utils.backcompat import experimental_arg_warning
 
-from .definition_config_schema import IDefinitionConfigSchema
+from .definition_config_schema import (
+    IDefinitionConfigSchema,
+    convert_user_facing_definition_config_schema,
+)
 from .hook_definition import HookDefinition
 from .inference import infer_output_props
 from .input import In, InputDefinition
 from .output import Out, OutputDefinition
-from .solid_definition import SolidDefinition
 
 if TYPE_CHECKING:
+    from dagster._core.definitions.asset_layer import AssetLayer
+
     from .composition import PendingNodeInvocation
     from .decorators.solid_decorator import DecoratedSolidFunction
 
 OpComputeFunction: TypeAlias = Callable[..., Any]
 
 
-class OpDefinition(SolidDefinition):
+class OpDefinition(NodeDefinition):
     """
     Defines an op, the functional unit of user-defined computation.
 
@@ -87,6 +107,12 @@ class OpDefinition(SolidDefinition):
             )
     """
 
+    _compute_fn: Union[Callable[..., Any], "DecoratedSolidFunction"]
+    _config_schema: IDefinitionConfigSchema
+    _required_resource_keys: AbstractSet[str]
+    _version: Optional[str]
+    _retry_policy: Optional[RetryPolicy]
+
     def __init__(
         self,
         compute_fn: Union[Callable[..., Any], "DecoratedSolidFunction"],
@@ -119,8 +145,10 @@ class OpDefinition(SolidDefinition):
                 explicit_input_defs=input_defs,
                 exclude_nothing=True,
             )
+            self._compute_fn = compute_fn
         else:
             resolved_input_defs = input_defs
+            self._compute_fn = check.callable_param(compute_fn, "compute_fn")
 
         code_version = canonicalize_backcompat_args(
             code_version, "code_version", version, "version", "2.0"
@@ -131,17 +159,28 @@ class OpDefinition(SolidDefinition):
             compute_fn=compute_fn, outs=outs, default_code_version=code_version
         )
 
+        self._config_schema = convert_user_facing_definition_config_schema(config_schema)
+        self._required_resource_keys = frozenset(
+            check.opt_set_param(required_resource_keys, "required_resource_keys", of_type=str)
+        )
+        self._version = check.opt_str_param(version, "version")
+        if version:
+            experimental_arg_warning("version", "OpDefinition.__init__")
+        self._retry_policy = check.opt_inst_param(retry_policy, "retry_policy", RetryPolicy)
+
+        positional_inputs = (
+            self._compute_fn.positional_inputs()
+            if isinstance(self._compute_fn, DecoratedSolidFunction)
+            else None
+        )
+
         super(OpDefinition, self).__init__(
-            compute_fn=compute_fn,
             name=name,
+            input_defs=check.sequence_param(resolved_input_defs, "input_defs", InputDefinition),
+            output_defs=check.sequence_param(output_defs, "output_defs", OutputDefinition),
             description=description,
-            config_schema=config_schema,
-            required_resource_keys=required_resource_keys,
-            tags=tags,
-            version=code_version,
-            retry_policy=retry_policy,
-            input_defs=resolved_input_defs,
-            output_defs=output_defs,
+            tags=check.opt_mapping_param(tags, "tags", key_type=str),
+            positional_inputs=positional_inputs,
         )
 
     @property
@@ -154,6 +193,11 @@ class OpDefinition(SolidDefinition):
 
     @public  # type: ignore
     @property
+    def name(self) -> str:
+        return super(OpDefinition, self).name
+
+    @public  # type: ignore
+    @property
     def ins(self) -> Mapping[str, In]:
         return {input_def.name: In.from_definition(input_def) for input_def in self.input_defs}
 
@@ -162,26 +206,30 @@ class OpDefinition(SolidDefinition):
     def outs(self) -> Mapping[str, Out]:
         return {output_def.name: Out.from_definition(output_def) for output_def in self.output_defs}
 
+    @property
+    def compute_fn(self) -> Union[Callable[..., Any], "DecoratedSolidFunction"]:
+        return self._compute_fn
+
+    @public  # type: ignore
+    @property
+    def config_schema(self) -> IDefinitionConfigSchema:
+        return self._config_schema
+
     @public  # type: ignore
     @property
     def required_resource_keys(self) -> AbstractSet[str]:
-        return super(OpDefinition, self).required_resource_keys
+        return frozenset(self._required_resource_keys)
 
     @public  # type: ignore
     @property
     def version(self) -> Optional[str]:
         deprecation_warning("`version` property on OpDefinition", "2.0")
-        return super(OpDefinition, self).version
+        return self._version
 
     @public  # type: ignore
     @property
     def retry_policy(self) -> Optional[RetryPolicy]:
-        return super(OpDefinition, self).retry_policy
-
-    @public  # type: ignore
-    @property
-    def name(self) -> str:
-        return super(OpDefinition, self).name
+        return self._retry_policy
 
     @public  # type: ignore
     @property
@@ -203,6 +251,194 @@ class OpDefinition(SolidDefinition):
     @public
     def with_retry_policy(self, retry_policy: RetryPolicy) -> "PendingNodeInvocation":
         return super(OpDefinition, self).with_retry_policy(retry_policy)
+
+    def is_from_decorator(self) -> bool:
+        from .decorators.solid_decorator import DecoratedSolidFunction
+
+        return isinstance(self._compute_fn, DecoratedSolidFunction)
+
+    def get_output_annotation(self) -> Any:
+        if not self.is_from_decorator():
+            raise DagsterInvalidInvocationError(
+                f"Attempted to get output annotation for {self.node_type_str} '{self.name}', "
+                "which was not constructed from a decorated function."
+            )
+        return cast("DecoratedSolidFunction", self.compute_fn).get_output_annotation()
+
+    def all_dagster_types(self) -> Iterator[DagsterType]:
+        yield from self.all_input_output_types()
+
+    def iterate_node_defs(self) -> Iterator[NodeDefinition]:
+        yield self
+
+    def iterate_solid_defs(self) -> Iterator["OpDefinition"]:
+        yield self
+
+    T_Handle = TypeVar("T_Handle", bound=Optional[NodeHandle])
+
+    def resolve_output_to_origin(
+        self, output_name: str, handle: T_Handle
+    ) -> Tuple[OutputDefinition, T_Handle]:
+        return self.output_def_named(output_name), handle
+
+    def resolve_output_to_origin_op_def(self, output_name: str) -> "OpDefinition":
+        return self
+
+    def get_inputs_must_be_resolved_top_level(
+        self, asset_layer: "AssetLayer", handle: Optional[NodeHandle] = None
+    ) -> Sequence[InputDefinition]:
+        handle = cast(NodeHandle, check.inst_param(handle, "handle", NodeHandle))
+        unresolveable_input_defs = []
+        for input_def in self.input_defs:
+            if (
+                not input_def.dagster_type.loader
+                and not input_def.dagster_type.kind == DagsterTypeKind.NOTHING
+                and not input_def.root_manager_key
+                and not input_def.has_default_value
+            ):
+                input_asset_key = asset_layer.asset_key_for_input(handle, input_def.name)
+                # If input_asset_key is present, this input can be resolved
+                # by a source asset, so input does not need to be resolved
+                # at the top level.
+                if input_asset_key:
+                    continue
+                unresolveable_input_defs.append(input_def)
+        return unresolveable_input_defs
+
+    def input_has_default(self, input_name: str) -> bool:
+        return self.input_def_named(input_name).has_default_value
+
+    def default_value_for_input(self, input_name: str) -> InputDefinition:
+        return self.input_def_named(input_name).default_value
+
+    def input_supports_dynamic_output_dep(self, input_name: str) -> bool:
+        return True
+
+    def copy_for_configured(
+        self,
+        name: str,
+        description: Optional[str],
+        config_schema: IDefinitionConfigSchema,
+        config_or_config_fn: Any,
+    ) -> "OpDefinition":
+        return OpDefinition(
+            name=name,
+            ins={name: In.from_definition(input_def) for input_def in self.input_defs},
+            outs={name: Out.from_definition(output_def) for output_def in self.output_defs},
+            compute_fn=self.compute_fn,
+            config_schema=config_schema,
+            description=description or self.description,
+            tags=self.tags,
+            required_resource_keys=self.required_resource_keys,
+            version=self.version,
+            retry_policy=self.retry_policy,
+        )
+
+    def get_resource_requirements(
+        self,
+        outer_context: Optional[object] = None,
+    ) -> Iterator[ResourceRequirement]:
+        # Outer requiree in this context is the outer-calling node handle. If not provided, then just use the solid name.
+        outer_context = cast(Optional[Tuple[NodeHandle, Optional["AssetLayer"]]], outer_context)
+        if not outer_context:
+            handle = None
+            asset_layer = None
+        else:
+            handle, asset_layer = outer_context
+        node_description = f"{self.node_type_str} '{handle or self.name}'"
+        for resource_key in sorted(list(self.required_resource_keys)):
+            yield OpDefinitionResourceRequirement(
+                key=resource_key, node_description=node_description
+            )
+        for input_def in self.input_defs:
+            if input_def.root_manager_key:
+                yield InputManagerRequirement(
+                    key=input_def.root_manager_key,
+                    node_description=node_description,
+                    input_name=input_def.name,
+                    root_input=True,
+                )
+            elif input_def.input_manager_key:
+                yield InputManagerRequirement(
+                    key=input_def.input_manager_key,
+                    node_description=node_description,
+                    input_name=input_def.name,
+                    root_input=False,
+                )
+            elif asset_layer and handle:
+                input_asset_key = asset_layer.asset_key_for_input(handle, input_def.name)
+                if input_asset_key:
+                    io_manager_key = asset_layer.io_manager_key_for_asset(input_asset_key)
+                    yield InputManagerRequirement(
+                        key=io_manager_key,
+                        node_description=node_description,
+                        input_name=input_def.name,
+                        root_input=False,
+                    )
+
+        for output_def in self.output_defs:
+            yield OutputManagerRequirement(
+                key=output_def.io_manager_key,
+                node_description=node_description,
+                output_name=output_def.name,
+            )
+
+    def __call__(self, *args, **kwargs) -> Any:
+        from ..execution.context.invocation import UnboundSolidExecutionContext
+        from .composition import is_in_composition
+        from .decorators.solid_decorator import DecoratedSolidFunction
+
+        if is_in_composition():
+            return super(OpDefinition, self).__call__(*args, **kwargs)
+        else:
+            node_label = self.node_type_str  # string "solid" for solids, "op" for ops
+
+            if not isinstance(self.compute_fn, DecoratedSolidFunction):
+                raise DagsterInvalidInvocationError(
+                    f"Attemped to invoke {node_label} that was not constructed using the `@{node_label}` "
+                    f"decorator. Only {node_label}s constructed using the `@{node_label}` decorator can be "
+                    "directly invoked."
+                )
+            if self.compute_fn.has_context_arg():
+                if len(args) + len(kwargs) == 0:
+                    raise DagsterInvalidInvocationError(
+                        f"Compute function of {node_label} '{self.name}' has context argument, but no context "
+                        "was provided when invoking."
+                    )
+                if len(args) > 0:
+                    if args[0] is not None and not isinstance(
+                        args[0], UnboundSolidExecutionContext
+                    ):
+                        raise DagsterInvalidInvocationError(
+                            f"Compute function of {node_label} '{self.name}' has context argument, "
+                            "but no context was provided when invoking."
+                        )
+                    context = args[0]
+                    return solid_invocation_result(self, context, *args[1:], **kwargs)
+                # Context argument is provided under kwargs
+                else:
+                    context_param_name = get_function_params(self.compute_fn.decorated_fn)[0].name
+                    if context_param_name not in kwargs:
+                        raise DagsterInvalidInvocationError(
+                            f"Compute function of {node_label} '{self.name}' has context argument "
+                            f"'{context_param_name}', but no value for '{context_param_name}' was "
+                            f"found when invoking. Provided kwargs: {kwargs}"
+                        )
+                    context = kwargs[context_param_name]
+                    kwargs_sans_context = {
+                        kwarg: val
+                        for kwarg, val in kwargs.items()
+                        if not kwarg == context_param_name
+                    }
+                    return solid_invocation_result(self, context, *args, **kwargs_sans_context)
+
+            else:
+                if len(args) > 0 and isinstance(args[0], UnboundSolidExecutionContext):
+                    raise DagsterInvalidInvocationError(
+                        f"Compute function of {node_label} '{self.name}' has no context argument, but "
+                        "context was provided when invoking."
+                    )
+                return solid_invocation_result(self, None, *args, **kwargs)
 
 
 def _resolve_output_defs_from_outs(
