+++ python_modules/libraries/dagster-airflow/dagster_airflow/dagster_pipeline_factory.py
@@ -18,6 +18,7 @@ from airflow.models.baseoperator import BaseOperator
 from airflow.models.dag import DAG
 from airflow.models.dagbag import DagBag
 from airflow.settings import LOG_FORMAT
+from dagster._core.definitions.op_definition import OpDefinition
 from dagster_airflow.patch_airflow_example_dag import patch_airflow_example_dag
 
 from dagster import (
@@ -35,7 +36,7 @@ from dagster import _check as check
 from dagster import op, repository, resource
 from dagster._core.definitions.utils import VALID_NAME_REGEX, validate_tags
 from dagster._core.instance import AIRFLOW_EXECUTION_DATE_STR, IS_AIRFLOW_INGEST_PIPELINE_STR
-from dagster._legacy import ModeDefinition, PipelineDefinition, SolidDefinition
+from dagster._legacy import ModeDefinition, PipelineDefinition
 from dagster._utils.schedules import is_valid_cron_schedule
 
 # pylint: disable=no-name-in-module,import-error
@@ -586,7 +587,7 @@ def _traverse_airflow_dag(
     check.inst_param(dag, "dag", DAG)
     check.inst_param(task, "task", BaseOperator)
     check.list_param(seen_tasks, "seen_tasks", BaseOperator)
-    check.list_param(solid_defs, "solid_defs", SolidDefinition)
+    check.list_param(solid_defs, "solid_defs", OpDefinition)
     check.bool_param(use_airflow_template_context, "use_airflow_template_context")
     unique_id = check.opt_int_param(unique_id, "unique_id")
     mock_xcom = check.opt_bool_param(mock_xcom, "mock_xcom")
