+++ python_modules/libraries/dagstermill/dagstermill_tests/test_context.py
@@ -1,8 +1,8 @@
 from dagstermill.manager import MANAGER_FOR_NOTEBOOK_INSTANCE
 
 from dagster._core.definitions.dependency import Node
+from dagster._core.definitions.op_definition import OpDefinition
 from dagster._core.system_config.objects import ResolvedRunConfig
-from dagster._legacy import SolidDefinition
 
 BARE_OUT_OF_PIPELINE_CONTEXT = MANAGER_FOR_NOTEBOOK_INSTANCE.get_context()
 
@@ -45,7 +45,7 @@ def test_resources():
 
 
 def test_solid_def():
-    assert isinstance(BARE_OUT_OF_PIPELINE_CONTEXT.solid_def, SolidDefinition)
+    assert isinstance(BARE_OUT_OF_PIPELINE_CONTEXT.solid_def, OpDefinition)
 
 
 def test_solid():
